+++ pintree/src/main/java/com/trio/pintree/login/properties/KaKaoOauthProperties.java
@@ -1,4 +1,4 @@
-package com.trio.pintree.login.component;
+package com.trio.pintree.login.properties;
 
 import lombok.Getter;
 import lombok.Setter;
