+++ pintree/src/main/java/com/trio/pintree/login/service/KakaoOauthService.java
@@ -1,6 +1,6 @@
 package com.trio.pintree.login.service;
 
-import com.trio.pintree.login.component.KaKaoOauthProperties;
+import com.trio.pintree.login.properties.KaKaoOauthProperties;
 import com.trio.pintree.login.domain.Member;
 import com.trio.pintree.login.dto.AccessTokenResponse;
 import com.trio.pintree.login.dto.KaKaoAccessTokenResponse;
