+++ pintree/src/main/java/com/trio/pintree/login/service/OauthServiceFactory.java
@@ -1,4 +1,4 @@
-package com.trio.pintree.login.component;
+package com.trio.pintree.login.service;
 
 import com.trio.pintree.login.dto.AccessTokenResponse;
 import com.trio.pintree.login.service.OauthService;
